Implement HKDF, build system, directory structure
This is the Dafny implementation of the HKDF as described in (https://tools.ietf.org/html/rfc5869), based on an initial version by Rustan Leino. The HKDF relies on an HMAC provided by a library. `CryptoMac.dfy` contains extern methods that are implemented in `hkdf-extern.cs` as wrappers around BouncyCastle. Supports SHA256 and SHA384 hash functions.

To make it testable I created a different directory structure and set up a lit configuration. The tests currently added are the HKDF test vectors as in github.com/aws/aws-encryption-sdk-c.
The negative test case (test vector 6), when no hash algorithm is chosen, is not yet implemented.