Implements HKDF spec, HMac wrapper, std library
The HKDF spec provides mathematical functions implementing the HKDF's expand step using sequences. These will later be used in the implementation of the imperative version of the HKDF as pre- and post-conditions and loop invariants.

The HMac wrapper mirrors the C#/Java interface of the BouncyCastle library.

StandardLibrary is extended with functionality needed for the implementation/verification of the HKDF.