Implements the HKDF, Array lib, test vectors, C# wrappers
Implementation of the HKDF that uses the HKDF spec commited previously.

Array library wrapper.

HKDF test vectors taken from https://github.com/aws/aws-encryption-sdk-c. The negative test case which does not use a hash function (test vector 6) is not implemented as we require that either SHA256 or SHA384 are used as a pre-condition.

We connect the Dafny implementation with BouncyCastle using the wrappers in `HKDF-extern.cs`.