Implements the message header deserialization
Implements https://code.amazon.com/packages/TmpAWSEncryptionSDKSpec/blobs/mainline/--/message-header.md

There are some TODOs left in the code. Especially for parts of the code which are weakly specified or not yet specified, or for which I could not find the specification. The message header especially is not clearly specified yet.