Factored uint into a submodule of the standard library
This is a refactoring affecting many lines, but it is not large conceptually. I am moving the uint definitions and some, partially new, helper functions into their own submodule `StandardLibrary.Uint`. Furthermore I am moving the StandardLibrary to its own folder. Along with this I renamed `byte` to `uint8` for consistency with the other uint types.

Implement `uint16` to array and to seq and back

Implement `uint32` to array and to seq and back

Add functional tests for uint conversions