Removing unused files from merge for now
memcmp_le_none in Sort.dfy was causing verification errors from time to time. It was used in the Antisymmetry lemma, which however was not used in EncCtx, suggesting that a total order was not required. Anyhow, as we do not know the full spec yet, it seems that these files need to be re-engineered at some point and it is not worth keeping them in the merge as they are disruptive.

They should be merged again when they're up to spec.