Make ToyClient resolve, verify, compile, and run
Update Makefile to include all the files needed.
Update Makefile to split build and verify targets (which makes it easy to rebuild without verifying everything).
Rewrite Main to only use the RSA keyring (not a multi-keyring with both RSA and AES).
Reintroduc `EC_PUBLIC_KEY_FIELD` constant in Materials.dfy.
Mark some recursive routines in StandardLibrary with `{:opaque}`, which will improve verifier performance.

Note, although I did fix up most of the formatting, I did not change all variables names according to our conventions.

We still need to compare the specs and code to the English specification. This PR is not intended to address the various Issues we have recorded as TODOs.