Compute message signature in footer
The signature is computed as a pair of BigInteger's, (r, s), which is then converted into
a byte sequence. The byte sequence is expected to be of a particular length, according
to the ECDSA parameter. If (r, s) does not generate a byte sequence of the expected
length, then the code instead converts (r, N-s) to a byte sequence, where N is the order
of the group. If this byte sequence isn't of the right length, either, then the code gives
up trying.

This is similar to what ESDK Java and ESDK JavaScript do. However, those implementations
never give up. Instead, if neither (r, s) nor (r, N-s) produces a byte sequence of the
right length, then the process computes a new signature (r, s) and repeats. Possibly forever.

Maybe the code should retry a few times and then give up?