Address further PR comments
Rename ID.Algorithm() to ID.EncryptionSuite().
Change behavior to allow final frame to be as long as the frame length, instead of producing a 0-length final frame.
Make `EncryptRegularFrame`'s `length` parameter ghost, to allow it to be used in the method specification without needing to be passed at run time.
Add TODO in Streams.dfy.