Add caching CMM
This caching CMM does not share the Cryptographic Materials Cache. Therefore,
it does not make use of a partition ID.

Change the specifications of the CMM's GetEncryptionMaterials and DecryptMaterials
methods to say that Repr may change. This change had some rippling consequences in
Client, ToyClient, and Main.

Add postcondition `Valid() ==> this in Repr` to CMM..Valid(). This is a good idea for
validity predicates.

Add a predicate ContainsIdentityKDF to AlgorithmSuite.ID.

Add a utility module Time, which has a method that returns the current time in seconds
since the Unix epoch.