Respond to PR comments
I added a test routine that checks the postcondition of `Signature.KeyGen`.
In particular, this checks that the public key returned is y-compressed,
as indicated by the first byte being either 2 or 3.

I pulled the key serialization into a separate method. Such a method could be written
in Dafny, so that we can reuse it across implementations. However, it needs to
invoke `ECPoint` methods, which are specific to C#. Therefore, I wrote this key serialization
method in C#.

I added a comment in the code to refer to the relevant section in www.secg.org/sec1-v2.pdf.

Regarding the desired to have `xBytes` be zero-padded, I am unable to find any documentation
regarding either an `.FPCurve.Q` or a `GetBytesCount(true)`. I tried many variations of
both, including `.Parameters.Q` and just `.Q` for the former, but to no avail. Experimentally,
however, I consistently get the same length for `xBytes` (namely, 48 when I'm using
`ECDSA_P384`).

I added Issue https://github.com/awslabs/aws-encryption-sdk-dafny/issues/130, which says to
get the ESDK to reject verification keys that do not compress y.