Attempt to avoid time-out
The y-compression PR caused an unexpected verification error on the build server in some code that
was not touched. The source location reported as what the verifier was working on when it timed out
(which is not certain to be accurate) revealed an unexpected subset-type check. This commit removes
the need for that subset-type check by tightening up the result type of Deserialize.InsertNewEntry.
All along, everything verified on my local machine, and the hope is that this change may help the
verification on the build server as well.