Y compression (#105)
* Compress y coordinate in signature key generation

* Specify Signature.KeyGen to use y-compression

Add postcondition to Signature.KeyGen to say that public has uses y-compression. Since
this is an extern method, it would be nice to do further checks of this postcondition in some
sanity tests for externs.

* Respond to PR comments

I added a test routine that checks the postcondition of `Signature.KeyGen`.
In particular, this checks that the public key returned is y-compressed,
as indicated by the first byte being either 2 or 3.

I pulled the key serialization into a separate method. Such a method could be written
in Dafny, so that we can reuse it across implementations. However, it needs to
invoke `ECPoint` methods, which are specific to C#. Therefore, I wrote this key serialization
method in C#.

I added a comment in the code to refer to the relevant section in www.secg.org/sec1-v2.pdf.

Regarding the desired to have `xBytes` be zero-padded, I am unable to find any documentation
regarding either an `.FPCurve.Q` or a `GetBytesCount(true)`. I tried many variations of
both, including `.Parameters.Q` and just `.Q` for the former, but to no avail. Experimentally,
however, I consistently get the same length for `xBytes` (namely, 48 when I'm using
`ECDSA_P384`).

I added Issue https://github.com/awslabs/aws-encryption-sdk-dafny/issues/130, which says to
get the ESDK to reject verification keys that do not compress y.

* Zero-pad encoding of x

* Add comment about expecting a prime curve

The comment states a precondition of `SerializePublicKey`. The implementation of
`SerializePublicKey` throws an exception if the precondition is not met. The one caller
of `SerializePublicKey` gets its curve from the name "secp256r1" or "secp384r1", and
thus the curve ought to be a prime curve.

* Attempt to avoid time-out

The y-compression PR caused an unexpected verification error on the build server in some code that
was not touched. The source location reported as what the verifier was working on when it timed out
(which is not certain to be accurate) revealed an unexpected subset-type check. This commit removes
the need for that subset-type check by tightening up the result type of Deserialize.InsertNewEntry.
All along, everything verified on my local machine, and the hope is that this change may help the
verification on the build server as well.