fix: Indicate different nativeType for JavaScript (#224)
* fix: Indicate different nativeType for JavaScript

Dafny detects that newtype StrengthBits can be compiled to an unsigned integer, because
it contains no negative values. Unfortunately, the extern call in .NET that uses StrengthBits takes
a signed integer. Fortunately, the code can use a :nativeType attribute on StrengthBit to change
the target type to "int". Unfortunately, this does not work for JavaScript, which only supports
"number" as a native integer type. Fortunately, Dafny recently started supporting :nativeType
with a list of target types, which means that both "int" and "number" can be listed
(https://github.com/dafny-lang/dafny/pull/556).

The following was a previous attempt to solve the problem for JavaScript, but it causes the
reverse problem for .NET:
Closes #78

There are at least two other problems with building for JavaScript at this time. One is that
the Dafny-to-JavaScript compiler treats an :extern module as something that also has a
native definition. The other is some malformed JavaScript produced by that compiler. These
issues will be addressed separately.

* Use latest version of Dafny, which supports new :nativeType