chore: Remove redundant postconditions
Many of the postconditions of `GetEncryptionMaterials` and `DecryptMaterials`
followed from the validity of the output parameter. In turn, the validity of the
output parameter follows from its `Valid...` type.

In the case of `GetEncryptionMaterials`, two of the postconditions were
conveniently captured into a new member function `EncryptionMaterials.Serializable`
(which is a predicate that is also needed in the upcoming Caching CMM).