Use dafny 3.1 (#364)
This PR brings the source code up-to-date for Dafny 3.1. Type checking and verification pass.

Additionally, this PR removes trailing spaces in the source code.

robin-aws: I've also switched the CodeBuild spec to download a Dafny release rather than building from source, which wasn't working any more on the CodeBuild images. We should be able to fast-follow with upgrading to 3.2.

See also #367 - it was necessary to temporarily remove definiteAssignment:3 to allow one particularly tricky proof to verify. We need to prioritize fixing this too.