Replace while loop over set with compiled exists expression (#526)
The old version of this method was using the `:|` idiom for looping over a set, which takes quadratic time because it involves creating N copies of the set as it is whittled down.

The new version simply uses the exact `exists` expression used in the specification. It may not have been done this way in the first place because it's not obvious that such expressions can be compiled!