test(NET): NativeWrappers validate output.
Adds negative tests for the NativeWrapper output
validation logic.
However, tests are still WIP and have revealed that
there is something wrong with TypeConversion's
`ToDafny_CommonError_AwsCryptographicMaterialProvidersBaseException`.

Technically... addressing that is scope creep... so I am
going to create another SIM and move on to testing the
other NativeWrapper parts.