feat(NET): NativeWrappers use AwsCryptographicMaterialProvidersException
To circumnavigate the exception conversion issue, I refactored
the NativeWrappers to avoid the BaseException type.

I then refactored the tests to validate that the exception
message is preserved going to Dafny.

Finally, I manually refactored the output validation for the Keyring
NativeWrapper so that we can better describe the exception.