WIP feat(Exception): Convert any Exception to ServiceBase Exception
To ensure that our library ALWAYS returns a Dafny _IResult when
calling back from native code, we must convert every exception.

This change is only applied to Cryptographic Material Providers,
but will be applied to the ESDK as well.

This PR is the "intent" for current WIP in the Polymorph code generation
package.